Altered locations of imports and some tests to allow for travis integration
