+++ .travis.yml
@@ -1,9 +1,9 @@
 language: python
 python:
   - "2.7"
-
+cache: pip
 install:
   - sudo apt-get update
-  - wget https://raw.githubusercontent.com/cctbx/cctbx_project/master/libtbx/auto_build/bootstrap.py
-  - python ./bootstrap.py
-  - cctbx-python setup.py install
+  - python2.7 -m pip install -r requirements.txt
+script:
+  - python2.7 scripts/run_tests.py
