+++ requirements.txt
@@ -1,0 +1,5 @@
+enum34
+numpy
+pandas
+pyjob ==0.2.1
+gemmi
