+++ simbad/util/pdb_util.py
@@ -1,6 +1,4 @@
 import gemmi
-import iotbx.pdb
-import iotbx.pdb.fetch
 import logging
 import numpy as np
 import os
@@ -37,6 +35,7 @@ class PdbStructure(object):
 
     @staticmethod
     def get_pdb_content(pdb_code):
+        import iotbx.pdb.fetch
         try:
             try:
                 content = iotbx.pdb.fetch.fetch(pdb_code, data_type="pdb", format="pdb", mirror="pdb-redo")
